A formula for the terminal Space Invaders clone posted on HN.
HN article: https://news.ycombinator.com/item?id=10078928
Written in Golang, this is a completely ASCII art implementation of
Space Invaders which works in the terminal.  Compatible with any POSIX
compliant system that can run Go.

Since my original work, I have updated this formula to version 1.1 of
the game and have renamed it from spaceinvaders to spaceinvaders-go to
prevent naming conflicts as requested in PR #305.  This is a rebase of 4
commits to squash for nice history.